+++ src/main/java/com/revlearn/team1/dto/course/request/CourseStudentDTO.java
@@ -1,4 +1,4 @@
-package com.revlearn.team1.dto.request;
+package com.revlearn.team1.dto.course.request;
 
 public record CourseStudentDTO(
         Long courseId,
