+++ src/main/java/com/revlearn/team1/dto/course/response/CourseEducatorResDTO.java
@@ -1,4 +1,4 @@
-package com.revlearn.team1.dto.response;
+package com.revlearn.team1.dto.course.response;
 
 public record CourseEducatorResDTO(
         String message,
