+++ src/main/java/com/revlearn/team1/dto/course/response/CourseStudentResDTO.java
@@ -1,4 +1,4 @@
-package com.revlearn.team1.dto.response;
+package com.revlearn.team1.dto.course.response;
 
 public record CourseStudentResDTO(
         String message,
