+++ src/main/java/com/revlearn/team1/service/CourseService.java
@@ -1,10 +1,10 @@
 package com.revlearn.team1.service;
 
-import com.revlearn.team1.dto.CourseDTO;
-import com.revlearn.team1.dto.request.CourseEducatorDTO;
-import com.revlearn.team1.dto.request.CourseStudentDTO;
-import com.revlearn.team1.dto.response.CourseEducatorResDTO;
-import com.revlearn.team1.dto.response.CourseStudentResDTO;
+import com.revlearn.team1.dto.course.CourseDTO;
+import com.revlearn.team1.dto.course.request.CourseEducatorDTO;
+import com.revlearn.team1.dto.course.request.CourseStudentDTO;
+import com.revlearn.team1.dto.course.response.CourseEducatorResDTO;
+import com.revlearn.team1.dto.course.response.CourseStudentResDTO;
 
 import java.util.List;
