+++ src/test/java/com/revlearn/team1/unit/service/CourseServiceImpTest.java
@@ -1,8 +1,8 @@
 package com.revlearn.team1.unit.service;
 
-import com.revlearn.team1.dto.CourseDTO;
-import com.revlearn.team1.dto.request.CourseEducatorDTO;
-import com.revlearn.team1.dto.response.CourseEducatorResDTO;
+import com.revlearn.team1.dto.course.CourseDTO;
+import com.revlearn.team1.dto.course.request.CourseEducatorDTO;
+import com.revlearn.team1.dto.course.response.CourseEducatorResDTO;
 import com.revlearn.team1.enums.AttendanceMethod;
 import com.revlearn.team1.exceptions.CourseNotFoundException;
 import com.revlearn.team1.mapper.CourseMapper;
